move _shuffle_context
