+++ python/dask_cudf/dask_cudf/sorting.py
@@ -6,7 +6,6 @@ import cupy
 import numpy as np
 import tlz as toolz
 
-from dask import config
 from dask.base import tokenize
 from dask.dataframe import methods
 from dask.dataframe.core import DataFrame, Index, Series
@@ -297,26 +296,3 @@ def sort_values(
         df4.divisions = tuple(methods.tolist(divisions))
 
     return df4
-
-
-def _shuffle_context(func, *args, shuffle="tasks", **kwargs):
-    # Utility to check `shuffle` kwarg and use dask-config
-    # context for "explicit-comms" option, when necessary
-    if shuffle not in ("tasks", "explicit-comms"):
-        raise ValueError(
-            f"Dask-cudf only supports in-memory shuffling with "
-            f"'tasks' or 'explicit-comms'. Got shuffle={shuffle}"
-        )
-    if shuffle == "explicit-comms":
-        try:
-            import dask_cuda  # noqa: F401
-        except ImportError:
-            raise ValueError(
-                "shuffle='explicit-comms' requires dask_cuda. "
-                "Please install dask_cuda, or use shuffle='tasks'."
-            )
-        # Use explicit-comms shuffle
-        with config.set({"explicit-comms": True}):
-            return func(*args, shuffle="tasks", **kwargs)
-
-    return func(*args, shuffle=shuffle, **kwargs)
